+++ backend/internal/autosnapshot/daemon.go
@@ -20,7 +20,7 @@ type Repo interface {
 	GetActiveConnectionCount(ctx context.Context, clusterName, instanceName string) (int, error)
 	GetBlockedSessionCount(ctx context.Context, clusterName, instanceName, databaseName string) (int, error)
 	GetInstanceInfo(ctx context.Context, clusterName, instanceName string) (dto.InstanceInfo, error)
-	GetQueriesReport(ctx context.Context, clusterName, instanceName string, excludeUsers []string) ([]dto.QueryReport, error)
+	GetQueriesReport(ctx context.Context, clusterName, instanceName, databaseName string, excludeUsers []string) ([]dto.QueryReport, error)
 	GetQueriesBlocked(ctx context.Context, clusterName, instanceName, databaseName string) ([]dto.QueryBlocked, error)
 	GetPgssStatsResetTime(ctx context.Context, clusterName, instanceName, databaseName string) (*dto.StatsResetTime, error)
 	ResetQueryStats(ctx context.Context, clusterName, instanceName, databaseName string) error